refactor: remove cycle from Compensation
