+++ src/components/Employee/Compensation/Compensation.tsx
@@ -3,7 +3,6 @@ import { useEffect, useMemo, useState } from 'react'
 import { Form } from 'react-aria-components'
 import { FormProvider, useForm, type SubmitHandler } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
-import * as v from 'valibot'
 import {
   useJobsAndCompensationsGetJobsSuspense,
   invalidateJobsAndCompensationsGetJobs,
@@ -16,23 +15,27 @@ import { useLocationsGetMinimumWagesSuspense } from '@gusto/embedded-api/react-q
 import { useEmployeeAddressesGetWorkAddressesSuspense } from '@gusto/embedded-api/react-query/employeeAddressesGetWorkAddresses'
 import { type Job } from '@gusto/embedded-api/models/components/job'
 import { useQueryClient } from '@gusto/embedded-api/ReactSDKProvider'
-import type { MinimumWage } from '@gusto/embedded-api/models/components/minimumwage'
 import type { FlsaStatusType } from '@gusto/embedded-api/models/components/flsastatustype'
 import { List } from './List'
 import { Head } from './Head'
 import { Edit } from './Edit'
 import { Actions } from './Actions'
+import {
+  type CompensationInputs,
+  type CompensationOutputs,
+  CompensationProvider,
+  CompensationSchema,
+  type MODE,
+} from './useCompensation'
 import type { RequireAtLeastOne } from '@/types/Helpers'
 import type { PAY_PERIODS } from '@/shared/constants'
-import { componentEvents, FLSA_OVERTIME_SALARY_LIMIT, FlsaStatus } from '@/shared/constants'
+import { componentEvents, FlsaStatus } from '@/shared/constants'
 import { useI18n } from '@/i18n'
-import { yearlyRate } from '@/helpers/payRateCalculator'
 import {
   BaseComponent,
   type BaseComponentInterface,
   useBase,
   type CommonComponentInterface,
-  createCompoundContext,
 } from '@/components/Base'
 import type { EmployeeOnboardingContextInterface } from '@/components/Flow/EmployeeOnboardingFlow'
 import { useFlow } from '@/components/Flow/Flow'
@@ -49,89 +52,6 @@ interface CompensationProps extends CommonComponentInterface {
   startDate: string
   defaultValues?: CompensationDefaultValues
 }
-type MODE =
-  | 'LIST'
-  | 'ADD_ADDITIONAL_JOB'
-  | 'ADD_INITIAL_JOB'
-  | 'EDIT_ADDITIONAL_JOB'
-  | 'EDIT_INITIAL_JOB'
-  | 'PROCEED'
-
-type CompensationContextType = {
-  employeeJobs: Job[]
-  currentJob?: Job | null
-  primaryFlsaStatus?: string
-  isPending: boolean
-  mode: MODE
-  showFlsaChangeWarning: boolean
-  minimumWages: MinimumWage[]
-  submitWithEffect: (newMode: MODE) => void
-  handleAdd: () => void
-  handleEdit: (uuid: string) => void
-  handleDelete: (uuid: string) => void
-  handleFlsaChange: (status: string | number) => void
-  handleCancelAddJob: () => void
-}
-const [useCompensation, CompensationProvider] =
-  createCompoundContext<CompensationContextType>('CompensationContext')
-export { useCompensation }
-
-const CompensationSchema = v.intersect([
-  v.object({
-    jobTitle: v.pipe(v.string(), v.nonEmpty()),
-  }),
-  v.variant('adjustForMinimumWage', [
-    v.object({
-      adjustForMinimumWage: v.literal(true),
-      minimumWageId: v.pipe(v.string(), v.nonEmpty()),
-    }),
-    v.object({ adjustForMinimumWage: v.literal(false) }),
-  ]),
-  v.variant('flsaStatus', [
-    v.pipe(
-      v.object({
-        flsaStatus: v.union([
-          v.literal(FlsaStatus.EXEMPT),
-          v.literal(FlsaStatus.SALARIED_NONEXEMPT),
-          v.literal(FlsaStatus.NONEXEMPT),
-        ]),
-        paymentUnit: v.union([
-          v.literal('Hour'),
-          v.literal('Week'),
-          v.literal('Month'),
-          v.literal('Year'),
-        ]),
-        rate: v.pipe(v.number(), v.minValue(1), v.transform(String)),
-      }),
-      //Exempt salary threshold validation:
-      v.forward(
-        v.check(input => {
-          return (
-            //TODO: this should not be validated for non-primary jobs for NONEXEMPT
-            input.flsaStatus !== FlsaStatus.EXEMPT ||
-            yearlyRate(Number(input.rate), input.paymentUnit) >= FLSA_OVERTIME_SALARY_LIMIT
-          )
-        }),
-        ['flsaStatus'],
-      ),
-    ),
-    v.object({
-      flsaStatus: v.literal(FlsaStatus.OWNER),
-      paymentUnit: v.literal('Paycheck'),
-      rate: v.pipe(v.number(), v.minValue(1), v.transform(String)),
-    }),
-    v.object({
-      flsaStatus: v.union([
-        v.literal(FlsaStatus.COMMISSION_ONLY_EXEMPT),
-        v.literal(FlsaStatus.COMISSION_ONLY_NONEXEMPT),
-      ]),
-      paymentUnit: v.literal('Year'),
-      rate: v.pipe(v.literal(0), v.transform(String)),
-    }),
-  ]),
-])
-export type CompensationInputs = v.InferInput<typeof CompensationSchema>
-export type CompensationOutputs = v.InferOutput<typeof CompensationSchema>
 
 export function Compensation(props: CompensationProps & BaseComponentInterface) {
   return (
